pyproject.toml
[tool.poetry]
name = "game-loader-to-clickhouse"
version = "0.1.0"
description = "Service for loading flattened CS2 game data from JSON into ClickHouse and triggered by RabbitMQ events."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false
exclude = [".venv"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pika = "^1.3.2"
pydantic = "^2.9.0"
pydantic-settings = "^2.2.1"
python-dateutil = "^2.9.0.post0"
clickhouse-connect = "^0.7.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.6.0"
isort = "^5.13.2"
autoflake = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = ["venv", ".venv", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.isort]
profile = "black"
line_length = 88
combine_as_imports = true
multi_line_output = 3
skip = [".venv", "venv", "__pycache__"]

[tool.pytest.ini_options]
addopts = "-v --disable-warnings"
testpaths = ["tests"]
asyncio_mode = "auto"
